All tests passing for session store and window store
