+++ streams/src/test/java/org/apache/kafka/streams/state/internals/RocksDBSessionStoreTest.java
@@ -16,337 +16,61 @@
  */
 package org.apache.kafka.streams.state.internals;
 
-import org.apache.kafka.common.metrics.Metrics;
-import org.apache.kafka.common.serialization.Serdes;
-import org.apache.kafka.common.utils.LogContext;
-import org.apache.kafka.streams.KeyValue;
+import java.util.HashSet;
+import org.apache.kafka.common.serialization.Serde;
 import org.apache.kafka.streams.kstream.Windowed;
 import org.apache.kafka.streams.kstream.internals.SessionWindow;
-import org.apache.kafka.streams.processor.internals.MockStreamsMetrics;
 import org.apache.kafka.streams.processor.internals.testutil.LogCaptureAppender;
 import org.apache.kafka.streams.state.KeyValueIterator;
 import org.apache.kafka.streams.state.SessionStore;
 import org.apache.kafka.streams.state.Stores;
-import org.apache.kafka.test.InternalMockProcessorContext;
-import org.apache.kafka.test.NoOpRecordCollector;
-import org.apache.kafka.test.TestUtils;
-import org.junit.After;
-import org.junit.Before;
 import org.junit.Test;
 
 import java.util.Arrays;
-import java.util.Collections;
-import java.util.List;
 
 import static java.time.Duration.ofMillis;
-import static org.apache.kafka.test.StreamsTestUtils.toList;
-import static org.apache.kafka.test.StreamsTestUtils.valuesToList;
-import static org.hamcrest.CoreMatchers.equalTo;
-import static org.hamcrest.CoreMatchers.hasItem;
-import static org.hamcrest.MatcherAssert.assertThat;
+import static org.apache.kafka.test.StreamsTestUtils.valuesToSet;
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertTrue;
 
-public class RocksDBSessionStoreTest {
+public class RocksDBSessionStoreTest extends SessionBytesStoreTest {
 
-    private SessionStore<String, Long> sessionStore;
-    private InternalMockProcessorContext context;
+    private static final String STORE_NAME = "rocksDB session store";
 
-    @Before
-    public void before() {
-        sessionStore = Stores.sessionStoreBuilder(
+    @Override
+    <K, V> SessionStore<K, V> buildSessionStore(final long retentionPeriod,
+                                                 final Serde<K> keySerde,
+                                                 final Serde<V> valueSerde) {
+        return Stores.sessionStoreBuilder(
             Stores.persistentSessionStore(
-                "session-store",
-                ofMillis(10_000L)),
-            Serdes.String(),
-            Serdes.Long()).build();
-
-        context = new InternalMockProcessorContext(
-            TestUtils.tempDirectory(),
-            Serdes.String(),
-            Serdes.Long(),
-            new NoOpRecordCollector(),
-            new ThreadCache(
-                new LogContext("testCache "),
-                0,
-                new MockStreamsMetrics(new Metrics())));
-
-        sessionStore.init(context, sessionStore);
+                STORE_NAME,
+                ofMillis(retentionPeriod)),
+            keySerde,
+            valueSerde).build();
     }
 
-    @After
-    public void close() {
-        sessionStore.close();
+    @Override
+    String getMetricsScope() {
+        return new RocksDbSessionBytesStoreSupplier(null, 0).metricsScope();
+    }
+
+    @Override
+    void setClassLoggerToDebug() {
+        LogCaptureAppender.setClassLoggerToDebug(AbstractRocksDBSegmentedBytesStore.class);
     }
 
     @Test
-    public void shouldPutAndFindSessionsInRange() {
-        final String key = "a";
-        final Windowed<String> a1 = new Windowed<>(key, new SessionWindow(10, 10L));
-        final Windowed<String> a2 = new Windowed<>(key, new SessionWindow(500L, 1000L));
-        sessionStore.put(a1, 1L);
-        sessionStore.put(a2, 2L);
-        sessionStore.put(new Windowed<>(key, new SessionWindow(1500L, 2000L)), 1L);
-        sessionStore.put(new Windowed<>(key, new SessionWindow(2500L, 3000L)), 2L);
-
-        final List<KeyValue<Windowed<String>, Long>> expected =
-            Arrays.asList(KeyValue.pair(a1, 1L), KeyValue.pair(a2, 2L));
-
-        try (final KeyValueIterator<Windowed<String>, Long> values =
-                 sessionStore.findSessions(key, 0, 1000L)
-        ) {
-            assertEquals(expected, toList(values));
-        }
-
-        final List<KeyValue<Windowed<String>, Long>> expected2 = Collections.singletonList(KeyValue.pair(a2, 2L));
-
-        try (final KeyValueIterator<Windowed<String>, Long> values2 =
-                 sessionStore.findSessions(key, 400L, 600L)
-        ) {
-            assertEquals(expected2, toList(values2));
-        }
-    }
-
-    @Test
-    public void shouldFetchAllSessionsWithSameRecordKey() {
-        final List<KeyValue<Windowed<String>, Long>> expected = Arrays.asList(
-            KeyValue.pair(new Windowed<>("a", new SessionWindow(0, 0)), 1L),
-            KeyValue.pair(new Windowed<>("a", new SessionWindow(10, 10)), 2L),
-            KeyValue.pair(new Windowed<>("a", new SessionWindow(100, 100)), 3L),
-            KeyValue.pair(new Windowed<>("a", new SessionWindow(1000, 1000)), 4L));
-
-        for (final KeyValue<Windowed<String>, Long> kv : expected) {
-            sessionStore.put(kv.key, kv.value);
-        }
-
-        // add one that shouldn't appear in the results
-        sessionStore.put(new Windowed<>("aa", new SessionWindow(0, 0)), 5L);
-
-        try (final KeyValueIterator<Windowed<String>, Long> values = sessionStore.fetch("a")) {
-            assertEquals(expected, toList(values));
-        }
-    }
-
-    @Test
-    public void shouldFetchAllSessionsWithinKeyRange() {
-        final List<KeyValue<Windowed<String>, Long>> expected = Arrays.asList(
-            KeyValue.pair(new Windowed<>("aa", new SessionWindow(10, 10)), 2L),
-            KeyValue.pair(new Windowed<>("aaa", new SessionWindow(100, 100)), 3L),
-            KeyValue.pair(new Windowed<>("b", new SessionWindow(1000, 1000)), 4L),
-            KeyValue.pair(new Windowed<>("bb", new SessionWindow(1500, 2000)), 5L));
-
-        for (final KeyValue<Windowed<String>, Long> kv : expected) {
-            sessionStore.put(kv.key, kv.value);
-        }
-
-        // add some that shouldn't appear in the results
+    public void shouldRemoveExpired() {
         sessionStore.put(new Windowed<>("a", new SessionWindow(0, 0)), 1L);
-        sessionStore.put(new Windowed<>("bbb", new SessionWindow(2500, 3000)), 6L);
+        sessionStore.put(new Windowed<>("aa", new SessionWindow(0, SEGMENT_INTERVAL)), 2L);
+        sessionStore.put(new Windowed<>("a", new SessionWindow(10, SEGMENT_INTERVAL)), 3L);
 
-        try (final KeyValueIterator<Windowed<String>, Long> values = sessionStore.fetch("aa", "bb")) {
-            assertEquals(expected, toList(values));
-        }
-    }
-
-    @Test
-    public void shouldFetchExactSession() {
-        sessionStore.put(new Windowed<>("a", new SessionWindow(0, 4)), 1L);
-        sessionStore.put(new Windowed<>("aa", new SessionWindow(0, 3)), 2L);
-        sessionStore.put(new Windowed<>("aa", new SessionWindow(0, 4)), 3L);
-        sessionStore.put(new Windowed<>("aa", new SessionWindow(1, 4)), 4L);
-        sessionStore.put(new Windowed<>("aaa", new SessionWindow(0, 4)), 5L);
-
-        final long result = sessionStore.fetchSession("aa", 0, 4);
-        assertEquals(3L, result);
-    }
-
-    @Test
-    public void shouldFindValuesWithinMergingSessionWindowRange() {
-        final String key = "a";
-        sessionStore.put(new Windowed<>(key, new SessionWindow(0L, 0L)), 1L);
-        sessionStore.put(new Windowed<>(key, new SessionWindow(1000L, 1000L)), 2L);
-
-        final List<KeyValue<Windowed<String>, Long>> expected = Arrays.asList(
-            KeyValue.pair(new Windowed<>(key, new SessionWindow(0L, 0L)), 1L),
-            KeyValue.pair(new Windowed<>(key, new SessionWindow(1000L, 1000L)), 2L));
-
-        try (final KeyValueIterator<Windowed<String>, Long> results =
-                 sessionStore.findSessions(key, -1, 1000L)) {
-            assertEquals(expected, toList(results));
-        }
-    }
-
-    @Test
-    public void shouldRemove() {
-        sessionStore.put(new Windowed<>("a", new SessionWindow(0, 1000)), 1L);
-        sessionStore.put(new Windowed<>("a", new SessionWindow(1500, 2500)), 2L);
-
-        sessionStore.remove(new Windowed<>("a", new SessionWindow(0, 1000)));
-
-        try (final KeyValueIterator<Windowed<String>, Long> results =
-                 sessionStore.findSessions("a", 0L, 1000L)) {
-            assertFalse(results.hasNext());
-        }
-
-        try (final KeyValueIterator<Windowed<String>, Long> results =
-                 sessionStore.findSessions("a", 1500L, 2500L)) {
-            assertTrue(results.hasNext());
-        }
-    }
-
-    @Test
-    public void shouldRemoveOnNullAggValue() {
-        sessionStore.put(new Windowed<>("a", new SessionWindow(0, 1000)), 1L);
-        sessionStore.put(new Windowed<>("a", new SessionWindow(1500, 2500)), 2L);
-
-        sessionStore.put(new Windowed<>("a", new SessionWindow(0, 1000)), null);
-
-        try (final KeyValueIterator<Windowed<String>, Long> results =
-            sessionStore.findSessions("a", 0L, 1000L)) {
-            assertFalse(results.hasNext());
-        }
-
-        try (final KeyValueIterator<Windowed<String>, Long> results =
-            sessionStore.findSessions("a", 1500L, 2500L)) {
-            assertTrue(results.hasNext());
-        }
-    }
-
-    @Test
-    public void shouldFindSessionsToMerge() {
-        final Windowed<String> session1 = new Windowed<>("a", new SessionWindow(0, 100));
-        final Windowed<String> session2 = new Windowed<>("a", new SessionWindow(101, 200));
-        final Windowed<String> session3 = new Windowed<>("a", new SessionWindow(201, 300));
-        final Windowed<String> session4 = new Windowed<>("a", new SessionWindow(301, 400));
-        final Windowed<String> session5 = new Windowed<>("a", new SessionWindow(401, 500));
-        sessionStore.put(session1, 1L);
-        sessionStore.put(session2, 2L);
-        sessionStore.put(session3, 3L);
-        sessionStore.put(session4, 4L);
-        sessionStore.put(session5, 5L);
-
-        try (final KeyValueIterator<Windowed<String>, Long> results =
-                 sessionStore.findSessions("a", 150, 300)
-        ) {
-            assertEquals(session2, results.next().key);
-            assertEquals(session3, results.next().key);
-            assertFalse(results.hasNext());
-        }
-    }
-
-    @Test
-    public void shouldFetchExactKeys() {
-        sessionStore = Stores.sessionStoreBuilder(
-            Stores.persistentSessionStore(
-                "session-store",
-                ofMillis(0x7a00000000000000L)),
-            Serdes.String(),
-            Serdes.Long()).build();
-
-        sessionStore.init(context, sessionStore);
-
-        sessionStore.put(new Windowed<>("a", new SessionWindow(0, 0)), 1L);
-        sessionStore.put(new Windowed<>("aa", new SessionWindow(0, 10)), 2L);
-        sessionStore.put(new Windowed<>("a", new SessionWindow(10, 20)), 3L);
-        sessionStore.put(new Windowed<>("aa", new SessionWindow(10, 20)), 4L);
-        sessionStore.put(new Windowed<>("a", new SessionWindow(0x7a00000000000000L - 2, 0x7a00000000000000L - 1)), 5L);
+        // Advance stream time to expire the first record
+        sessionStore.put(new Windowed<>("aa", new SessionWindow(10, 2 * SEGMENT_INTERVAL)), 4L);
 
         try (final KeyValueIterator<Windowed<String>, Long> iterator =
-                 sessionStore.findSessions("a", 0, Long.MAX_VALUE)
+            sessionStore.findSessions("a", "b", 0L, Long.MAX_VALUE)
         ) {
-            assertThat(valuesToList(iterator), equalTo(Arrays.asList(1L, 3L, 5L)));
+            assertEquals(valuesToSet(iterator), new HashSet<>(Arrays.asList(2L, 3L, 4L)));
         }
-
-        try (final KeyValueIterator<Windowed<String>, Long> iterator =
-                 sessionStore.findSessions("aa", 0, Long.MAX_VALUE)
-        ) {
-            assertThat(valuesToList(iterator), equalTo(Arrays.asList(2L, 4L)));
-        }
-
-        try (final KeyValueIterator<Windowed<String>, Long> iterator =
-                 sessionStore.findSessions("a", "aa", 0, Long.MAX_VALUE)
-        ) {
-            assertThat(valuesToList(iterator), equalTo(Arrays.asList(1L, 3L, 2L, 4L, 5L)));
-        }
-
-        try (final KeyValueIterator<Windowed<String>, Long> iterator =
-                 sessionStore.findSessions("a", "aa", 10, 0)
-        ) {
-            assertThat(valuesToList(iterator), equalTo(Collections.singletonList(2L)));
-        }
-    }
-
-    @Test
-    public void shouldReturnSameResultsForSingleKeyFindSessionsAndEqualKeyRangeFindSessions() {
-        sessionStore.put(new Windowed<>("a", new SessionWindow(0, 1)), 0L);
-        sessionStore.put(new Windowed<>("aa", new SessionWindow(2, 3)), 1L);
-        sessionStore.put(new Windowed<>("aa", new SessionWindow(4, 5)), 2L);
-        sessionStore.put(new Windowed<>("aaa", new SessionWindow(6, 7)), 3L);
-
-        final KeyValueIterator<Windowed<String>, Long> singleKeyIterator = sessionStore.findSessions("aa", 0L, 10L);
-        final KeyValueIterator<Windowed<String>, Long> keyRangeIterator = sessionStore.findSessions("aa", "aa", 0L, 10L);
-
-        assertEquals(singleKeyIterator.next(), keyRangeIterator.next());
-        assertEquals(singleKeyIterator.next(), keyRangeIterator.next());
-        assertFalse(singleKeyIterator.hasNext());
-        assertFalse(keyRangeIterator.hasNext());
-    }
-
-    @Test(expected = NullPointerException.class)
-    public void shouldThrowNullPointerExceptionOnFindSessionsNullKey() {
-        sessionStore.findSessions(null, 1L, 2L);
-    }
-
-    @Test(expected = NullPointerException.class)
-    public void shouldThrowNullPointerExceptionOnFindSessionsNullFromKey() {
-        sessionStore.findSessions(null, "anyKeyTo", 1L, 2L);
-    }
-
-    @Test(expected = NullPointerException.class)
-    public void shouldThrowNullPointerExceptionOnFindSessionsNullToKey() {
-        sessionStore.findSessions("anyKeyFrom", null, 1L, 2L);
-    }
-
-    @Test(expected = NullPointerException.class)
-    public void shouldThrowNullPointerExceptionOnFetchNullFromKey() {
-        sessionStore.fetch(null, "anyToKey");
-    }
-
-    @Test(expected = NullPointerException.class)
-    public void shouldThrowNullPointerExceptionOnFetchNullToKey() {
-        sessionStore.fetch("anyFromKey", null);
-    }
-
-    @Test(expected = NullPointerException.class)
-    public void shouldThrowNullPointerExceptionOnFetchNullKey() {
-        sessionStore.fetch(null);
-    }
-
-    @Test(expected = NullPointerException.class)
-    public void shouldThrowNullPointerExceptionOnRemoveNullKey() {
-        sessionStore.remove(null);
-    }
-
-    @Test(expected = NullPointerException.class)
-    public void shouldThrowNullPointerExceptionOnPutNullKey() {
-        sessionStore.put(null, 1L);
-    }
-
-    @Test
-    public void shouldNotThrowInvalidRangeExceptionWithNegativeFromKey() {
-        LogCaptureAppender.setClassLoggerToDebug(InMemoryWindowStore.class);
-        final LogCaptureAppender appender = LogCaptureAppender.createAndRegister();
-
-        final String keyFrom = Serdes.String().deserializer().deserialize("", Serdes.Integer().serializer().serialize("", -1));
-        final String keyTo = Serdes.String().deserializer().deserialize("", Serdes.Integer().serializer().serialize("", 1));
-
-        final KeyValueIterator<Windowed<String>, Long> iterator = sessionStore.findSessions(keyFrom, keyTo, 0L, 10L);
-        assertFalse(iterator.hasNext());
-
-        final List<String> messages = appender.getMessages();
-        assertThat(messages, hasItem("Returning empty iterator for fetch with invalid key range: from > to. "
-            + "This may be due to serdes that don't preserve ordering when lexicographically comparing the serialized bytes. "
-            + "Note that the built-in numerical serdes do not follow this for negative numbers"));
     }
 }
